rust: narrow openssl split for codacy gate
