+++ clients/rust/crates/rubin-consensus/src/verify_sig_openssl/digest.rs
@@ -1,4 +1,3 @@
-use super::ffi::{self, MdCtx, VerificationPkey};
 use crate::error::{ErrorCode, TxError};
 use core::ffi::CStr;
 
@@ -24,38 +23,7 @@ pub(crate) fn openssl_verify_sig_digest_oneshot(
     if pubkey.is_empty() || signature.is_empty() || msg.is_empty() {
         return Err(TxError::new(ErrorCode::TxErrParse, "openssl: empty input"));
     }
-
-    // SAFETY: all pointers passed to OpenSSL are derived from live Rust slices
-    // for the duration of the call; RAII wrappers own and free the created
-    // OpenSSL key/context exactly once.
-    unsafe {
-        openssl_sys::ERR_clear_error();
-        let pkey = VerificationPkey::new_raw_public_key(alg, pubkey)?;
-        let mctx = MdCtx::new("openssl: EVP_MD_CTX_new failed")?;
-        if ffi::EVP_DigestVerifyInit_ex(
-            mctx.as_mut_ptr(),
-            core::ptr::null_mut(),
-            core::ptr::null(),
-            core::ptr::null_mut(),
-            core::ptr::null(),
-            pkey.as_mut_ptr(),
-            core::ptr::null(),
-        ) <= 0
-        {
-            return Err(TxError::new(
-                ErrorCode::TxErrParse,
-                "openssl: EVP_DigestVerifyInit_ex failed",
-            ));
-        }
-        let rc = ffi::EVP_DigestVerify(
-            mctx.as_mut_ptr(),
-            signature.as_ptr(),
-            signature.len(),
-            msg.as_ptr(),
-            msg.len(),
-        );
-        map_digest_verify_rc(rc)
-    }
+    super::openssl_verify_sig_digest_oneshot_raw(alg, pubkey, signature, msg)
 }
 
 #[cfg(test)]
